refactor: comment dto패키지 이동
