+++ backend/src/main/java/com/woowacourse/gongseek/comment/application/CommentService.java
@@ -6,13 +6,13 @@ import com.woowacourse.gongseek.article.exception.ArticleNotFoundException;
 import com.woowacourse.gongseek.auth.application.dto.AppMember;
 import com.woowacourse.gongseek.auth.exception.NotAuthorException;
 import com.woowacourse.gongseek.auth.exception.NotMemberException;
+import com.woowacourse.gongseek.comment.application.dto.CommentRequest;
+import com.woowacourse.gongseek.comment.application.dto.CommentResponse;
+import com.woowacourse.gongseek.comment.application.dto.CommentUpdateRequest;
+import com.woowacourse.gongseek.comment.application.dto.CommentsResponse;
 import com.woowacourse.gongseek.comment.domain.Comment;
 import com.woowacourse.gongseek.comment.domain.repository.CommentRepository;
 import com.woowacourse.gongseek.comment.exception.CommentNotFoundException;
-import com.woowacourse.gongseek.comment.presentation.dto.CommentRequest;
-import com.woowacourse.gongseek.comment.presentation.dto.CommentResponse;
-import com.woowacourse.gongseek.comment.presentation.dto.CommentUpdateRequest;
-import com.woowacourse.gongseek.comment.presentation.dto.CommentsResponse;
 import com.woowacourse.gongseek.member.domain.Member;
 import com.woowacourse.gongseek.member.domain.repository.MemberRepository;
 import com.woowacourse.gongseek.member.exception.MemberNotFoundException;
