+++ backend/src/main/java/com/woowacourse/gongseek/comment/application/dto/CommentResponse.java
@@ -1,4 +1,4 @@
-package com.woowacourse.gongseek.comment.presentation.dto;
+package com.woowacourse.gongseek.comment.application.dto;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import com.woowacourse.gongseek.comment.domain.Comment;
