+++ backend/src/main/java/com/woowacourse/gongseek/comment/presentation/CommentController.java
@@ -3,9 +3,9 @@ package com.woowacourse.gongseek.comment.presentation;
 import com.woowacourse.gongseek.auth.application.dto.AppMember;
 import com.woowacourse.gongseek.auth.presentation.AuthenticationPrinciple;
 import com.woowacourse.gongseek.comment.application.CommentService;
-import com.woowacourse.gongseek.comment.presentation.dto.CommentRequest;
-import com.woowacourse.gongseek.comment.presentation.dto.CommentUpdateRequest;
-import com.woowacourse.gongseek.comment.presentation.dto.CommentsResponse;
+import com.woowacourse.gongseek.comment.application.dto.CommentRequest;
+import com.woowacourse.gongseek.comment.application.dto.CommentUpdateRequest;
+import com.woowacourse.gongseek.comment.application.dto.CommentsResponse;
 import javax.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
