+++ backend/src/test/java/com/woowacourse/gongseek/member/application/MemberServiceTest.java
@@ -12,9 +12,9 @@ import com.woowacourse.gongseek.article.domain.repository.ArticleRepository;
 import com.woowacourse.gongseek.auth.application.dto.GuestMember;
 import com.woowacourse.gongseek.auth.application.dto.LoginMember;
 import com.woowacourse.gongseek.comment.application.CommentService;
+import com.woowacourse.gongseek.comment.application.dto.CommentRequest;
 import com.woowacourse.gongseek.comment.domain.Comment;
 import com.woowacourse.gongseek.comment.domain.repository.CommentRepository;
-import com.woowacourse.gongseek.comment.presentation.dto.CommentRequest;
 import com.woowacourse.gongseek.member.domain.Member;
 import com.woowacourse.gongseek.member.domain.repository.MemberRepository;
 import com.woowacourse.gongseek.member.exception.MemberNotFoundException;
